agents/agentic-ui/ui-libs/streamlit/streamlit-langgraph-poc/requirements.txt
streamlit
langgraph 
langchain 
langchain-openai
python-dotenv
watchdog